requirements.txt
pytest==7.2.0
pytest-cov==4.0.0
python-dotenv==0.21.0
PyYAML==6.0